Started setting up dev dependencies for making tests. ServerJS now only starts the server and appJS is where development of server is made.
