+++ package.json
@@ -9,7 +9,8 @@
     "build": "vite build",
     "preview": "vite preview",
     "prepare": "husky install",
-    "start": "node src/server/server.js"
+    "start": "node src/server/server.js",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "husky": {
     "hooks": {
@@ -43,6 +44,8 @@
     "ejs": "^3.1.10",
     "express": "^4.21.2",
     "express-ejs-layouts": "^2.5.1",
-    "marked": "^15.0.6"
+    "jest": "^29.7.0",
+    "marked": "^15.0.6",
+    "supertest": "^7.0.0"
   }
 }
